Give access to double precision Constraint1DOF angle to C++ code

+++ Source/AGXUnrealBarrier/Public/Constraints/Constraint1DOFBarrier.h
@@ -31,7 +31,7 @@ public:
 	 * centimeters.
 	 * @return The current angle of the free degree of freedom.
 	 */
-	float GetAngle() const;
+	double GetAngle() const;
 
 	/**
 	 * This methods return the current speed for the 1D constraint.
